requirements.txt
streamlit>=1.40.1
google-generativeai==0.3.2
python-dotenv==1.0.0
